Move texture replacer to GPU/Common.
Add new files ReplacedTexture.cpp/h too.
