+++ GPU/Common/TextureReplacer.h
